Fix computation of download estimate.
The previous estimate assumed the last 100 samples were 100 blocks.

+++ src/qt/modaloverlay.h
@@ -46,7 +46,7 @@ private:
     Ui::ModalOverlay *ui;
     int bestHeaderHeight; //best known height (based on the headers)
     QDateTime bestHeaderDate;
-    QVector<QPair<qint64, double> > blockProcessTime;
+    QVector<QPair<qint64, int> > blockProcessTime;
     bool layerIsVisible;
     bool userClosed;
 };
